requirements.txt
absl-py==0.9.0
astor==0.8.1
atari-py==0.2.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.2.2
cmake==3.15.3
cycler==0.10.0
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.2
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
gym==0.15.4
h5py==2.10.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
numpy==1.17.4
oauthlib==3.1.0
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyglet==1.3.2
pyparsing==2.4.5
python-dateutil==2.8.1
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.0
six==1.13.0
tensorboard==2.0.2
tensorflow-estimator==2.0.1
termcolor==1.1.0
https://download.pytorch.org/whl/cu90/torch-1.0.1-cp37-cp37m-win_amd64.whl
torchvision==0.4.1
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
